Support nested glob patterns
Globbing separately for path/to/lib and path/to/lib/test would result in
trying to use intermediates/path/to/lib as both a file list file for the
first glob and as a directory to store the file list file for the second
glob, causing ninja to exit with an mkdir "Not a directory" error.

Add ".glob" to the file list file name to avoid collisions.

Change-Id: I8b9de9133d7ef667088a2f13eb28652508f3a290